+++ R/attrisk_analysis.R
@@ -62,7 +62,9 @@
 #'   contains the values \code{"Poor"} and \code{"Good"} for the first and
 #'   second levels, respectively, of each element in the \code{vars_response}
 #'   argument and that uses values in the \code{vars_response} argument as names
-#'   for the list.  The default value is NULL.
+#'   for the list.  If \code{response_levels} is provided without names,
+#'   then the names of \code{response_levels} are set to \code{vars_response}.
+#'   The default value is NULL.
 #'
 #' @param stressor_levels List providing the category values (levels) for each
 #'   element in the \code{vars_stressor} argument.  Each element in the list
@@ -74,7 +76,9 @@
 #'   contains the values \code{"Poor"} and \code{"Good"} for the first and
 #'   second levels, respectively, of each element in the \code{vars_stressor}
 #'   argument and that uses values in the \code{vars_stressor} argument as names
-#'   for the list.  The default value is NULL.
+#'   for the list.  If \code{stressor_levels} is provided without names,
+#'   then the names of \code{stressor_levels} are set to \code{vars_stressor}. 
+#'   The default value is NULL.
 #'
 #' @param subpops Vector composed of character values that identify the
 #'   names of subpopulation (domain) variables in \code{dframe}.
@@ -572,6 +576,9 @@ attrisk_analysis <- function(dframe, vars_response, vars_stressor, response_leve
         msg <- "Argument response_levels must be the same length as argument vars_response.\n"
         error_vec <- c(error_vec, msg)
       }
+      if (is.null(names(response_levels))) { # set default names if none provided
+        names(response_levels) <- vars_response
+      }
       if (any(sapply(response_levels, function(x) length(x) != 2))) {
         error_ind <- TRUE
         msg <- "Each element of argument response_levels must contain only two values.\n"
@@ -622,6 +629,9 @@ attrisk_analysis <- function(dframe, vars_response, vars_stressor, response_leve
         msg <- "Argument stressor_levels must be the same length as argument vars_stressor.\n"
         error_vec <- c(error_vec, msg)
       }
+      if (is.null(names(stressor_levels))) { # set default names if none provided
+        names(stressor_levels) <- vars_stressor
+      }
       if (any(sapply(stressor_levels, function(x) length(x) != 2))) {
         error_ind <- TRUE
         msg <- "Each element of argument stressor_levels must contain only two values.\n"
@@ -664,17 +674,17 @@ attrisk_analysis <- function(dframe, vars_response, vars_stressor, response_leve
   if (error_ind) {
     error_vec <<- error_vec
     if (length(error_vec) == 1) {
-      cat("During execution of the program, an error message was generated.  The error \nmessage is stored in a vector named 'error_vec'.  Enter the following command \nto view the error message: errorprnt()\n")
+      message("During execution of the program, an error message was generated.  The error \nmessage is stored in a vector named 'error_vec'.  Enter the following command \nto view the error message: errorprnt()\n")
     } else {
-      cat(paste("During execution of the program,", length(error_vec), "error messages were generated.  The error \nmessages are stored in a vector named 'error_vec'.  Enter the following \ncommand to view the error messages: errorprnt()\n"))
+      message(paste("During execution of the program,", length(error_vec), "error messages were generated.  The error \nmessages are stored in a vector named 'error_vec'.  Enter the following \ncommand to view the error messages: errorprnt()\n"))
     }
 
     if (warn_ind) {
       warn_df <<- warn_df
       if (nrow(warn_df) == 1) {
-        cat("During execution of the program, a warning message was generated.  The warning \nmessage is stored in a data frame named 'warn_df'.  Enter the following command \nto view the warning message: warnprnt()\n")
+        message("During execution of the program, a warning message was generated.  The warning \nmessage is stored in a data frame named 'warn_df'.  Enter the following command \nto view the warning message: warnprnt()\n")
       } else {
-        cat(paste("During execution of the program,", nrow(warn_df), "warning messages were generated.  The warning \nmessages are stored in a data frame named 'warn_df'.  Enter the following \ncommand to view the warning messages: warnprnt() \nTo view a subset of the warning messages (say, messages number 1, 3, and 5), \nenter the following command: warnprnt(m=c(1,3,5))\n"))
+        message(paste("During execution of the program,", nrow(warn_df), "warning messages were generated.  The warning \nmessages are stored in a data frame named 'warn_df'.  Enter the following \ncommand to view the warning messages: warnprnt() \nTo view a subset of the warning messages (say, messages number 1, 3, and 5), \nenter the following command: warnprnt(m=c(1,3,5))\n"))
       }
     }
     stop("See the preceding message(s).")
@@ -1218,9 +1228,9 @@ attrisk_analysis <- function(dframe, vars_response, vars_stressor, response_leve
   if (warn_ind) {
     warn_df <<- warn_df
     if (nrow(warn_df) == 1) {
-      cat("During execution of the program, a warning message was generated.  The warning \nmessage is stored in a data frame named 'warn_df'.  Enter the following command \nto view the warning message: warnprnt()\n")
+      message("During execution of the program, a warning message was generated.  The warning \nmessage is stored in a data frame named 'warn_df'.  Enter the following command \nto view the warning message: warnprnt()\n")
     } else {
-      cat(paste("During execution of the program,", nrow(warn_df), "warning messages were generated.  The warning \nmessages are stored in a data frame named 'warn_df'.  Enter the following \ncommand to view the warning messages: warnprnt() \nTo view a subset of the warning messages (say, messages number 1, 3, and 5), \nenter the following command: warnprnt(m=c(1,3,5))\n"))
+      message(paste("During execution of the program,", nrow(warn_df), "warning messages were generated.  The warning \nmessages are stored in a data frame named 'warn_df'.  Enter the following \ncommand to view the warning messages: warnprnt() \nTo view a subset of the warning messages (say, messages number 1, 3, and 5), \nenter the following command: warnprnt(m=c(1,3,5))\n"))
     }
   }