+++ tests/testthat/test-grts.R
@@ -238,6 +238,11 @@ if (on_solaris) {
       expect_equal(NCOL(grts_output$sites_near), 1)
     })
     
+    test_that("A warning (in message form) is produced", {
+      n_base <- c(low = 20, high = 30)
+      expect_message(expect_error(grts(NE_Lakes, n_base = n_base, stratum_var = "XYZ")))
+    })
+    
     #--------------------------------------
     #-------- Legacy
     #--------------------------------------